transport: merge transport-factory.h into transport-stack.h
These are interdependent interfaces and LogTransportStack will be the
sole user of LogTransportFactory.
